+++ src/traits/mod.rs
@@ -1,5 +1,6 @@
 pub mod executor;
 pub mod processor;
 
-pub use executor::{DagExecutor, ProcessorMap, DependencyGraph, EntryPoints};
+pub use executor::DagExecutor;
+pub use crate::config::{ProcessorMap, DependencyGraph, EntryPoints};
 pub use processor::Processor;
